pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-witness"
version = "0.1.0"
description = "Silent Observer agent architecture - separating awareness from thought for calibrated, trustworthy AI agency"
requires-python = ">=3.12,<3.15"
authors = [{ name = "Sterling Blood" }]

dependencies = [
    "google-genai",
    "huggingface-hub",
    "duckdb",
    "pydantic>=2.0",
    "pyyaml",
    "gradio>=4.0",
    "plotly",
    "opencv-python-headless",
    "numpy",
    "scipy",
    "sentence-transformers",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "pre-commit",
]

[project.scripts]
witness = "src.cockpit.app:main"

[tool.ruff]
target-version = "py312"
line-length = 99
lint.select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
